requirements.txt
# STB远程UI自动化测试项目依赖

# 基础依赖
pure-python-adb==0.3.0.dev0

# 图像识别依赖
opencv-python==4.8.1.78
pytesseract==0.3.10
Pillow==9.5.0
numpy==1.24.3

# 测试框架依赖
pytest==7.4.3
allure-pytest==2.13.2

# 测试报告生成依赖
jinja2==3.1.2
openpyxl==3.1.2

# 其他工具依赖
python-dotenv==1.0.0
requests==2.31.0